test_runner: escape tap in spec reporter
