+++ lib/internal/test_runner/reporter/spec.js
@@ -16,7 +16,7 @@ const Transform = require('internal/streams/transform');
 const { inspectWithNoCustomRetry } = require('internal/errors');
 const colors = require('internal/util/colors');
 const { kSubtestsFailed } = require('internal/test_runner/test');
-const { getCoverageReport } = require('internal/test_runner/utils');
+const { getCoverageReport, tapEscape } = require('internal/test_runner/utils');
 const { relative } = require('path');
 
 const symbols = {
@@ -73,12 +73,12 @@ class SpecReporter extends Transform {
     let symbol = symbols[type] ?? ' ';
     const { skip, todo } = data;
     const duration_ms = data.details?.duration_ms ? ` ${colors.gray}(${data.details.duration_ms}ms)${colors.white}` : '';
-    let title = `${data.name}${duration_ms}`;
+    let title = tapEscape(`${data.name}${duration_ms}`);
 
     if (skip !== undefined) {
-      title += ` # ${typeof skip === 'string' && skip.length ? skip : 'SKIP'}`;
+      title += ` # ${typeof skip === 'string' && skip.length ? tapEscape(skip) : 'SKIP'}`;
     } else if (todo !== undefined) {
-      title += ` # ${typeof todo === 'string' && todo.length ? todo : 'TODO'}`;
+      title += ` # ${typeof todo === 'string' && todo.length ? tapEscape(todo) : 'TODO'}`;
     }
     const error = this.#formatError(data.details?.error, indent);
     if (hasChildren) {
