+++ lib/internal/test_runner/reporter/tap.js
@@ -8,12 +8,11 @@ const {
   RegExpPrototypeSymbolSplit,
   SafeMap,
   SafeSet,
-  StringPrototypeReplaceAll,
   StringPrototypeRepeat,
 } = primordials;
 const { inspectWithNoCustomRetry } = require('internal/errors');
 const { isError, kEmptyObject } = require('internal/util');
-const { getCoverageReport } = require('internal/test_runner/utils');
+const { getCoverageReport, tapEscape } = require('internal/test_runner/utils');
 const kDefaultIndent = '    '; // 4 spaces
 const kFrameStartRegExp = /^ {4}at /;
 const kLineBreakRegExp = /\n|\r\n/;
@@ -108,20 +107,6 @@ function indent(nesting) {
   }
 
   return value;
-}
-
-
-// In certain places, # and \ need to be escaped as \# and \\.
-function tapEscape(input) {
-  let result = StringPrototypeReplaceAll(input, '\b', '\\b');
-  result = StringPrototypeReplaceAll(result, '\f', '\\f');
-  result = StringPrototypeReplaceAll(result, '\t', '\\t');
-  result = StringPrototypeReplaceAll(result, '\n', '\\n');
-  result = StringPrototypeReplaceAll(result, '\r', '\\r');
-  result = StringPrototypeReplaceAll(result, '\v', '\\v');
-  result = StringPrototypeReplaceAll(result, '\\', '\\\\');
-  result = StringPrototypeReplaceAll(result, '#', '\\#');
-  return result;
 }
 
 function jsToYaml(indent, name, value, seen) {
