+++ lib/internal/test_runner/utils.js
@@ -17,6 +17,7 @@ const {
   StringPrototypePadStart,
   StringPrototypePadEnd,
   StringPrototypeRepeat,
+  StringPrototypeReplaceAll,
   StringPrototypeSlice,
 } = primordials;
 
@@ -412,6 +413,19 @@ function getCoverageReport(pad, summary, symbol, color, table) {
   return report;
 }
 
+// In certain places, # and \ need to be escaped as \# and \\.
+function tapEscape(input) {
+  let result = StringPrototypeReplaceAll(input, '\b', '\\b');
+  result = StringPrototypeReplaceAll(result, '\f', '\\f');
+  result = StringPrototypeReplaceAll(result, '\t', '\\t');
+  result = StringPrototypeReplaceAll(result, '\n', '\\n');
+  result = StringPrototypeReplaceAll(result, '\r', '\\r');
+  result = StringPrototypeReplaceAll(result, '\v', '\\v');
+  result = StringPrototypeReplaceAll(result, '\\', '\\\\');
+  result = StringPrototypeReplaceAll(result, '#', '\\#');
+  return result;
+}
+
 module.exports = {
   convertStringToRegExp,
   countCompletedTest,
@@ -421,5 +435,6 @@ module.exports = {
   parseCommandLine,
   reporterScope,
   setupTestReporters,
+  tapEscape,
   getCoverageReport,
 };
